oauth refactor and response improvements
