+++ .env.example
@@ -1,5 +1,6 @@
 # The base URL the project will be served at
 PROXY_BASE_URL = http://127.0.0.1:8000
+APP_CLIENT_URL = http://127.0.0.1:5173
 
 # The connection URL so the timescaleDB instance can connect to the postgres database
 DATABASE_URL = postgresql://postgres:password@localhost:5432/tsdb
