+++ accounts/schemas.py
@@ -20,7 +20,7 @@ class OrganizationFields(Schema):
 class UserFields(Schema):
     first_name: str = Field(alias="firstName")
     last_name: str = Field(alias="lastName")
-    email: str
+    email: str = None
     middle_name: str = Field(default=None, alias="middleName")
     phone: str = None
     address: str = None
