+++ hydroserver/settings.py
@@ -145,6 +145,9 @@ AUTHLIB_OAUTH_CLIENTS = {
     }
 }
 
+APP_CLIENT_URL = config('APP_CLIENT_URL', default='http://127.0.0.1:5173')
+SECURE_CROSS_ORIGIN_OPENER_POLICY = None
+
 # Email Backend Settings
 
 EMAIL_BACKEND = 'django_ses.SESBackend'
